Move scenario_workers plugin to upgrades/conftest
