+++ tests/upgrades/conftest.py
@@ -86,12 +86,16 @@ import datetime
 import functools
 import json
 import os
+from pathlib import Path
 
 from box import Box
 import pytest
 
+from robottelo.config import configure_airgun, configure_nailgun, settings
 from robottelo.logging import logger
 from robottelo.utils.decorators.func_locker import lock_function
+
+json_file = Path('upgrade_workers.json')
 
 pre_upgrade_failed_tests = []
 
@@ -419,6 +423,57 @@ class DependentTestFailed(Exception):
     """Raise when the dependent test fails"""
 
     pass
+
+
+def save_worker_hostname(test_name, target_sat):
+    data = {}
+    if json_file.exists():
+        data = json.loads(json_file.read_text())
+    # Removing the parameter name from test name before save
+    test_name = test_name.split('[')[0] if '[' in test_name else test_name
+    data.update({test_name: target_sat.hostname})
+    json_file.write_text(json.dumps(data))
+
+
+@pytest.fixture(scope='session')
+def shared_workers():
+    if json_file.exists():
+        return json.loads(json_file.read_text())
+    return None
+
+
+def get_worker_hostname_from_testname(test_name, shared_workers):
+    return shared_workers.get(test_name)
+
+
+@pytest.fixture(autouse=True)
+def save_pre_upgrade_worker_hostname(request, target_sat):
+    """Saves the worker id of pre_upgrade test in json"""
+    if request.node.get_closest_marker('pre_upgrade'):
+        save_worker_hostname(request.node.name, target_sat)
+
+
+@pytest.fixture(autouse=True)
+def set_post_upgrade_hostname(request, shared_workers):
+    """Retrieves worker hostname of pre_upgrade and sets the same to run post_upgrade
+
+    The limitation of this implementation is that the XDIST_BEHAVIOR won't be observed,
+    and tests running under xdist will not be isolated.
+
+    This is overriding the session scoped align_to_satellite fixture
+    """
+    post_upgrade = request.node.get_closest_marker('post_upgrade')
+    if post_upgrade:
+        depend_test = post_upgrade.kwargs.get('depend_on')
+        if depend_test:
+            pre_test_name = depend_test.__name__
+            pre_hostname = get_worker_hostname_from_testname(pre_test_name, shared_workers)
+            if (pre_hostname is None) or (pre_hostname not in settings.server.hostnames):
+                pytest.skip(
+                    'Skipping the post_upgrade test as the pre_upgrade hostname was not found!'
+                )
+            configure_nailgun()
+            configure_airgun()
 
 
 @pytest.fixture(autouse=True)
